+++ src/Optimisarr.Api/Library/LibrarySeeder.cs
@@ -1,4 +1,5 @@
 using Microsoft.EntityFrameworkCore;
+using Optimisarr.Core.Domain;
 using Optimisarr.Data;
 
 namespace Optimisarr.Api.Library;